+++ libraries/compute-utils/model-catalog/src/model-handler.ts
@@ -4,7 +4,7 @@ import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 import {BehaviorSubject} from 'rxjs';
 import {filter, map, take} from 'rxjs/operators';
-import {getContextHelp, getPackage} from '../../shared-utils/utils';
+import {getContextHelp, getCurrentUserGroups, getPackage, requestGroupMembership} from '../../shared-utils/utils';
 
 function addPopover(popover: HTMLElement) {
   stylePopover(popover);
@@ -57,8 +57,7 @@ async function requestMembership(groupName: string) {
 
     const group = groups[0];
 
-    // Workaround till JS API is not ready: https://reddata.atlassian.net/browse/GROK-14160
-    await fetch(`${window.location.origin}/api/groups/${group.id}/requests/${grok.shell.user.group.id}`, {method: 'POST'});
+    await requestGroupMembership(group, grok.shell.user.group);
 
     grok.shell.info(`Request to join ${groupName} has been initiated. Please allow some time for approval.`);
   } catch (e: any) {
@@ -109,9 +108,7 @@ export class ModelHandler extends DG.ObjectHandler {
   }
 
   static async getUserGroups() {
-    // Workaround till JS API is not ready: https://reddata.atlassian.net/browse/GROK-14159
-    const userGroups = (await(await fetch(`${window.location.origin}/api/groups/all_parents`)).json() as DG.Group[]);
-    return userGroups;
+    return await getCurrentUserGroups();
   }
 
   static getMissingGroups(func: DG.Func, userGroups: DG.Group[]) {